fix: 컨벤션에 맞게 수정 및 로직 수정
