+++ src/main/java/spring/springserver/domain/community/comment/service/CommunityCommentServiceImpl.kt
@@ -18,7 +18,7 @@ import spring.springserver.global.exception.status_code.CommonStatusCode
 import java.time.LocalDateTime
 
 @Service
-@Transactional
+@Transactional(rollbackFor = [Exception::class])
 class CommunityCommentServiceImpl(private val communityCommentRepository: CommunityCommentRepository,
                                         private val communityCommentLikeRepository: CommunityCommentLikeRepository,
                                         private val communityAuthorizationService: CommunityAuthorizationService) : CommunityCommentService {
@@ -92,61 +92,5 @@ class CommunityCommentServiceImpl(private val communityCommentRepository: Commun
         communityComment.softDelete(LocalDateTime.now())
 
         return DeleteResponse.of("삭제되었습니다.")
-    }
-
-    override fun likeComment(communityCommentLikeRequest: CommunityCommentLikeRequest): CommunityLikeResponse {
-
-        val member = communityAuthorizationService.getCurrentMember()
-
-        val commentId = communityCommentLikeRequest.commentId
-
-        val communityComment = communityAuthorizationService.getActiveComment(commentId)
-
-        if (communityCommentLikeRepository.existsByMemberAndCommunityComment(member, communityComment)) {
-
-            throw ApplicationException.of(
-                CommonStatusCode.INVALID_ARGUMENT,
-                "이미 좋아요를 누른 댓글입니다."
-            )
-        }
-
-        communityCommentLikeRepository.save(
-            CommunityCommentLike(
-                member = member,
-                communityComment = communityComment,
-            )
-        )
-
-        return CommunityLikeResponse.of(
-            targetId = commentId,
-            likeCount = communityCommentLikeRepository.countByCommunityCommentId(commentId),
-            message = "댓글 좋아요가 등록되었습니다.",
-        )
-    }
-
-    override fun unlikeComment(commentId: Long): CommunityLikeResponse {
-
-        val member = communityAuthorizationService.getCurrentMember()
-
-        val communityComment = communityAuthorizationService.getActiveComment(commentId)
-
-        val deletedCount = communityCommentLikeRepository.deleteByMemberAndCommunityComment(
-            member,
-            communityComment
-        )
-
-        if (deletedCount == 0L) {
-
-            throw ApplicationException.of(
-                CommonStatusCode.INVALID_ARGUMENT,
-                "좋아요를 누르지 않은 댓글입니다."
-            )
-        }
-
-        return CommunityLikeResponse.of(
-            targetId = commentId,
-            likeCount = communityCommentLikeRepository.countByCommunityCommentId(commentId),
-            message = "댓글 좋아요가 취소되었습니다.",
-        )
     }
 }
