+++ src/main/java/spring/springserver/domain/community/post/data/response/CommunityPostResponse.kt
@@ -1,10 +1,11 @@
 package spring.springserver.domain.community.post.data.response
 
+import spring.springserver.domain.community.comment.repository.CommunityCommentRepository
 import spring.springserver.domain.community.post.entity.CommunityPost
 import java.time.LocalDateTime
 
 data class CommunityPostResponse(
-    val id: Long,
+    val id: Long?,
 
     val username: String,
 
@@ -24,10 +25,22 @@ data class CommunityPostResponse(
 ) {
     companion object {
 
+        fun toPostResponse(communityPost: CommunityPost,
+                           communityCommentRepository: CommunityCommentRepository): CommunityPostResponse {
+
+            val postId = communityPost.getId()!!
+
+            return of(
+                communityPost = communityPost,
+                commentCount = communityCommentRepository
+                    .countByCommunityPostIdAndDeletedAtIsNull(postId)
+            )
+        }
+
         fun of(communityPost: CommunityPost, commentCount: Long): CommunityPostResponse {
 
             return CommunityPostResponse(
-                communityPost.getId()!!,
+                communityPost.getId(),
                 communityPost.username,
                 communityPost.title,
                 communityPost.content,
